Further performance optimizations.
The big one is >>> 0 in peel, which ensures the numbers are positive.
Without this, accountForBias recursed way more than it was supposed to!

Also consoldiate int and integer, as I did when I proposed the change to core/Random.

Also use `rem` instead of `(%)` as it maps more closely to JS `%`.

That said, these optimizations don't improve things much (200ms MAX on the nonempty list suite)
but they open the door for hand-optimized bitwise ops.